+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 CHANGELOG.md
 
-v0.0.6a (29/10/2015)
+v0.0.8a (06/10/2015)
+- Created Properties(or Facilities) database collection
+- Connected Facilities view to Facilties database
+- Removed space at bottom of order card
+- Created Issues database collection and linked it to Work Orders page
+- Updated Facility card
+- Added functioning delete button to Facility and Work Order cards
+- Made Work Order description editable
+- Activated "new" button for Facilities and Work Orders
+
+v0.0.7a (05/10/2015)
+- Made header of order unselectable (to remove the "blue bar" on double click)
+- Reformatted work order form to better match screen designs
+- Made title field of issue editable
+
+v0.0.6a (29/09/2015)
 - Created detail view for a work order populated with an issued order
 - Removed dividers in left navigation menu
 - Left navigation menu now has fixed position
@@ -8,7 +23,7 @@ v0.0.6a (29/10/2015)
 - Fixed problem where clicking on body of expanded card would close the card
 - Issued and closed work orders identify contractor instead of requestor
 
-v0.0.5-a (28/10/2015)
+v0.0.5-a (28/09/2015)
 - Added Updated Work Order card placeholder to more closely resemble interface designs
 - Added Contract test data
 - Added Contract card / table row